# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VektraHub SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  client: 'client',
  admin: 'admin',
} as const;

export type BusinessPlan = typeof BusinessPlan[keyof typeof BusinessPlan];


export const BusinessPlan = {
  starter: 'starter',
  pro: 'pro',
} as const;

export type BusinessStatus = typeof BusinessStatus[keyof typeof BusinessStatus];


export const BusinessStatus = {
  active: 'active',
  suspended: 'suspended',
  trial: 'trial',
} as const;

export interface Business {
  id: number;
  name: string;
  sector: string;
  plan: BusinessPlan;
  status: BusinessStatus;
  createdAt?: string;
}

export interface UserProfile {
  clerkId: string;
  email: string;
  name?: string;
  role: UserProfileRole;
  business?: Business;
}

export interface UserProfileInput {
  name?: string;
}

export type OnboardInputSector = typeof OnboardInputSector[keyof typeof OnboardInputSector];


export const OnboardInputSector = {
  restaurante: 'restaurante',
  kiosco: 'kiosco',
  dietetica: 'dietetica',
  ferreteria: 'ferreteria',
  ropa: 'ropa',
  distribuidora: 'distribuidora',
  otro: 'otro',
} as const;

export type OnboardInputPlan = typeof OnboardInputPlan[keyof typeof OnboardInputPlan];


export const OnboardInputPlan = {
  starter: 'starter',
  pro: 'pro',
} as const;

export interface OnboardInput {
  businessName: string;
  sector: OnboardInputSector;
  plan: OnboardInputPlan;
}

export type BusinessAdminPlan = typeof BusinessAdminPlan[keyof typeof BusinessAdminPlan];


export const BusinessAdminPlan = {
  starter: 'starter',
  pro: 'pro',
} as const;

export type BusinessAdminStatus = typeof BusinessAdminStatus[keyof typeof BusinessAdminStatus];


export const BusinessAdminStatus = {
  active: 'active',
  suspended: 'suspended',
  trial: 'trial',
} as const;

export interface BusinessAdmin {
  id: number;
  name: string;
  sector: string;
  plan: BusinessAdminPlan;
  status: BusinessAdminStatus;
  ownerEmail: string;
  /** @nullable */
  ownerName?: string | null;
  productsCount: number;
  ordersCount: number;
  createdAt: string;
}

export type BusinessAdminUpdateStatus = typeof BusinessAdminUpdateStatus[keyof typeof BusinessAdminUpdateStatus];


export const BusinessAdminUpdateStatus = {
  active: 'active',
  suspended: 'suspended',
  trial: 'trial',
} as const;

export type BusinessAdminUpdatePlan = typeof BusinessAdminUpdatePlan[keyof typeof BusinessAdminUpdatePlan];


export const BusinessAdminUpdatePlan = {
  starter: 'starter',
  pro: 'pro',
} as const;

export interface BusinessAdminUpdate {
  status?: BusinessAdminUpdateStatus;
  plan?: BusinessAdminUpdatePlan;
}

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  stock?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  active: boolean;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  description?: string;
  price: number;
  category?: string;
  stock?: number;
  imageUrl?: string;
  active?: boolean;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  stock?: number;
  imageUrl?: string;
  active?: boolean;
}

export interface ProductStat {
  id: number;
  name: string;
  totalSold: number;
  revenue: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  items: OrderItem[];
  total: number;
  status: OrderStatus;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  customerPhone?: string;
  items: OrderItem[];
  notes?: string;
}

export type OrderUpdateStatus = typeof OrderUpdateStatus[keyof typeof OrderUpdateStatus];


export const OrderUpdateStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderUpdate {
  status?: OrderUpdateStatus;
  notes?: string;
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  totalOrders: number;
  totalSpent: number;
  createdAt: string;
}

export interface CustomerInput {
  name: string;
  email?: string;
  phone?: string;
}

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
}

export interface DashboardStats {
  totalOrders: number;
  totalRevenue: number;
  totalProducts: number;
  totalCustomers: number;
  ordersToday: number;
  revenueToday: number;
  pendingOrders: number;
  growthPercent: number;
}

export type ChatHistoryItemRole = typeof ChatHistoryItemRole[keyof typeof ChatHistoryItemRole];


export const ChatHistoryItemRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatHistoryItem {
  role: ChatHistoryItemRole;
  content: string;
}

export interface ChatMessageInput {
  message: string;
  history?: ChatHistoryItem[];
}

export interface ChatResponse {
  reply: string;
  suggestions?: string[];
}

export interface AdminStats {
  totalBusinesses: number;
  activeBusinesses: number;
  suspendedBusinesses: number;
  trialBusinesses: number;
  starterPlan: number;
  proPlan: number;
  newThisMonth: number;
}

export type ListProductsParams = {
category?: string;
search?: string;
};

export type ListOrdersParams = {
status?: ListOrdersStatus;
};

export type ListOrdersStatus = typeof ListOrdersStatus[keyof typeof ListOrdersStatus];


export const ListOrdersStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  preparing: 'preparing',
  ready: 'ready',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type ListCustomersParams = {
search?: string;
};

export type AdminListBusinessesParams = {
status?: AdminListBusinessesStatus;
};

export type AdminListBusinessesStatus = typeof AdminListBusinessesStatus[keyof typeof AdminListBusinessesStatus];


export const AdminListBusinessesStatus = {
  active: 'active',
  suspended: 'suspended',
  trial: 'trial',
} as const;