# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VektraHub SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for the current business
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, preparing, ready, delivered, cancelled]
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers for the current business
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/businesses:
    get:
      operationId: adminListBusinesses
      tags: [admin]
      summary: Admin - list all businesses
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, trial]
      responses:
        "200":
          description: All businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessAdmin"
  /admin/businesses/{id}:
    patch:
      operationId: adminUpdateBusiness
      tags: [admin]
      summary: Admin - update a business (activate, suspend, change plan)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessAdminUpdate"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessAdmin"
...
components:
  schemas:
...
    BusinessAdmin:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sector:
          type: string
        plan:
          type: string
          enum: [starter, pro]
        status:
          type: string
          enum: [active, suspended, trial]
        ownerEmail:
          type: string
        ownerName:
          type: ["string", "null"]
        productsCount:
          type: integer
        ordersCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sector
        - plan
        - status
        - ownerEmail
        - productsCount
        - ordersCount
        - createdAt
    BusinessAdminUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, trial]
        plan:
          type: string
          enum: [starter, pro]
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - items
        - total
        - status
        - createdAt
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
      required:
        - productId
        - productName
        - quantity
        - unitPrice
    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: string
      required:
        - customerName
        - items
    OrderUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        notes:
          type: string
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - totalOrders
        - totalSpent
        - createdAt
    CustomerInput:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
      required:
        - name
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
...