# lib/api-zod/src/generated/types/businessAdminUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VektraHub SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { BusinessAdminUpdatePlan } from './businessAdminUpdatePlan';
import type { BusinessAdminUpdateStatus } from './businessAdminUpdateStatus';

export interface BusinessAdminUpdate {
  status?: BusinessAdminUpdateStatus;
  plan?: BusinessAdminUpdatePlan;
}