# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VektraHub SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
}